// src/app/components/delivery-dashboard/model/placedLocation.model.ts
export interface PlacedLocationModel {
    orderId:     number;
    id:          number;
    addressName: string;
    userId:      number;
    phoneNumber: string;
    city:        string;
    zipCode:     number;
    details:     string;
    street:      string;
    lat:         number;
    lng:         number;
}